change "roll" to "rolling", add checks to flag
PH_US.m didnt properly handle modspec flags! Furthermore, codes were initially run on HPC servers with flag_rolling = "rolling"
but I subsequently changed it to "roll" (most likely because it's
shorter and confirms with the usage in the B-AR benchmarks models.
However, due to the fact that I have to rerun the US codes for the mod
specs other than rec, level, I have reverted back to the original in
both codes and added checks for the modspec flags