store all squared errors, not their mean
The changes in this and the subsequent commits on the recalc_RMSFE branch are in order to change the calculation of the RMSFE. From now on we calculate one RMSFE for each draw and then average. To implement these changes, the results_eval matfiles need to be adjusted as well as all subsequent eval functions that caclulate or use the RMSFE such as the tables. Furthermore, the cumsum function needs to be adjusted to take the mean of the squared errors before plotting